Ajuste em DTOs

+++ src/main/java/com/YEP4510/YEP4510/Repositories/ClubeRepository.java
@@ -4,5 +4,5 @@ import com.YEP4510.YEP4510.Models.Clube;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ClubeRepository extends CrudRepository<Clube, Integer>, PagingAndSortingRepository<Clube, Integer> {
+public interface ClubeRepository extends CrudRepository<Clube, Long>, PagingAndSortingRepository<Clube, Long> {
 }
